Pass integer to printf(3) precision.

+++ netcat.c
@@ -1587,7 +1587,7 @@ save_peer_cert(struct tls *tls_ctx, FILE *fp)
 
 	if ((pem = tls_peer_cert_chain_pem(tls_ctx, &plen)) == NULL)
 		errx(1, "Can't get peer certificate");
-	if (fprintf(fp, "%.*s", plen, pem) < 0)
+	if (fprintf(fp, "%.*s", (int)plen, pem) < 0)
 		err(1, "unable to save peer cert");
 	if (fflush(fp) != 0)
 		err(1, "unable to flush peer cert");
